% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_meta.R
\name{dictionary_meta}
\alias{dictionary_meta}
\title{Assess Dictionary Categories Within a Latent Semantic Space}
\usage{
dictionary_meta(dict, space = "auto", suggest = FALSE,
  suggestion_terms = 10, suggest_stopwords = FALSE,
  expand_cutoff_freq = 0.98, expand_cutoff_spaces = 10, glob = TRUE,
  space_dir = getOption("lingmatch.lspace.dir"), verbose = TRUE)
}
\arguments{
\item{dict}{A vector of terms, list of such vectors, or a matrix-like object to be
categorized by \code{\link{read.dic}}.}

\item{space}{A vector space used to calculate similarities between terms.
Name of a the space (see \code{\link{select.lspace}}), a matrix with terms as row names, or
\code{"auto"} to auto-select a space based on matched terms.}

\item{suggest}{Logical; if \code{TRUE}, will search for other terms for possible inclusion
in \code{space}.}

\item{suggestion_terms}{Number of terms to use when selecting suggested additions.}

\item{suggest_stopwords}{Logical; if \code{TRUE}, will potentially suggest function words.}

\item{expand_cutoff_freq}{Proportion of mapped terms to include when expanding dictionary terms.
Applies when \code{space} is a character (referring to a space to be loaded).}

\item{expand_cutoff_spaces}{Number of spaces in which a term has to appear to be considered
for expansion. Applies when \code{space} is a character (referring to a space to be loaded).}

\item{glob}{Logical; if \code{TRUE}, converts globs (asterisk wildcards) to regular expressions.}

\item{space_dir}{Directory from which \code{space} should be loaded.}

\item{verbose}{Logical; if \code{FALSE}, will not show status messages.}
}
\value{
A list
}
\description{
Assess Dictionary Categories Within a Latent Semantic Space
}
