% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lma_dict}
\alias{lma_dict}
\title{English function word category lists}
\usage{
lma_dict(..., as.regex = TRUE)
}
\arguments{
\item{...}{Numbers or letters corresponding to category names: ppron, ipron, article,
adverb, conj, prep, auxverb, negate, quant, interrog, number, or special.}

\item{as.regex}{Logical: if \code{FALSE}, lists are returned without regular expression.}
}
\description{
Returns a list of function words based on the Linguistic Inquiry and Word Count 2015 dictionary.
}
\note{
The \code{special} category is not returned unless specifically requested. It is a list of regular expression
strings attempting to capture special things like ellipses and emojis. If \code{special} is part of the returned list,
\code{as.regex} is set to \code{TRUE}.

The \code{special} list is always used by both \code{\link{lma_dtm}} and \code{\link{lma_termcat}}. When creating a dtm,
\code{special} is used to clean the original imput (so that, by default, the punctuation involed in ellipses and emojis
are treated as different -- as ellipses and emojis rather than as periods and parens and colons and such). When categorizing
a dtm, the input dictionary is passed by the special lists to be sure the terms in the dtm match up with the dictionary
(so, for example, ": (" would be replaced with "FROWN" in both the text and dictionary).
}
\examples{
#return the full dictionary (excluding special)
lma_dict()

#return the standard 7 category lsm categories
lma_dict(1:7)

#return just a few categories without regular expression
lma_dict(neg, int, aux, as.regex=FALSE)

#return special specifically
lma_dict(special)
}
