% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lma_meta}
\alias{lma_meta}
\title{Calculate Text-Based Metastatistics}
\usage{
lma_meta(text)
}
\arguments{
\item{text}{A character vector of texts.
\code{'document'} (default), \code{'sentence'}, or \code{'clause'}.}
}
\value{
A data.frame: \tabular{ll}{
  \code{characters} \tab Total number of characters.\cr
  \code{syllables} \tab Total number of syllables, as estimated by split lenght of
    \code{'a+[eu]*|e+a*|i+|o+[ui]*|u+|y+[aeiou]*'} - 1.\cr
  \code{words} \tab Total number of words (raw word count).\cr
  \code{unique_words} \tab Number of unique words (binary word count).\cr
  \code{clauses} \tab Number of clauses, as marked by commas, colons, semicolons, dashes, or brackets
    within sentences.\cr
  \code{sentences} \tab Number of sentences, as marked by periods, question marks, exclamation points,
    or new line characters.\cr
  \code{words_per_clause} \tab Average number of words per clause.\cr
  \code{words_per_sentence} \tab Average number of words per sentence.\cr
  \code{sixltr} \tab Number of words 6 or more characters long.\cr
  \code{characters_per_word} \tab Average number of characters per word
    (\code{characters} / \code{words}).\cr
  \code{syllables_per_word} \tab Average number of syllables per word
    (\code{syllables} / \code{words}).\cr
  \code{type_token_ratio} \tab Ratio of unique to total words: \code{unique_words} / \code{words}.\cr
  \code{reading_grade} \tab Flesch-Kincaid grade level: .39 * \code{words} / \code{sentences} +
    11.8 * \code{syllables} / \code{words} - 15.59, with \code{words} / \code{clause} or \code{words}
    in place of \code{words} / \code{sentences}, depending on \code{output}.\cr
  \code{numbers} \tab Number of terms starting with numbers. \cr
  \code{punct} \tab Number of terms starting with non-alphanumeric characters.\cr
  \code{periods} \tab Number of periods.\cr
  \code{commas} \tab Number of commas.\cr
  \code{qmarks} \tab Number of question marks.\cr
  \code{exclams} \tab Number of exclamation points.\cr
  \code{quotes} \tab Number of quotation marks (single and double).\cr
  \code{brackets} \tab Number of bracketing characters (including parentheses, and square,
    curly, and angle brackets).\cr
  \code{organizemarks} \tab Number of characters used for organization or structuring (including
    dashes, foreward slashes, colons, and semicolons).
}
}
\description{
Calculate simple descriptive statistics from text.
}
