% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingmatch.R
\name{lma_simets}
\alias{lma_simets}
\title{Similarity, Distance, and/or Accommodation Metrics}
\usage{
lma_simets(a, b = NULL, metric, metric.arg = list(), group = NULL,
  agg = TRUE, agg.mean = TRUE, square = TRUE, mean = FALSE,
  ncores = detectCores() - 2)
}
\arguments{
\item{a}{vector or matrix. If a vector, \code{b} must also be provided. If a matrix and \code{b}
is missing, each row will be compared. If a matrix and \code{b} is not missing, each row will
be compared with \code{b}.}

\item{b}{vector to be compared with \code{a} or rows of \code{a}.}

\item{metric}{a function (see details), or a character at least partially matching one of the
available metrics:
\tabular{ll}{
  \code{euclidean} \tab \code{1 / (1 + sum((a - b)^2)^.5)} \cr
  \code{canberra} \tab \code{mean(as.numeric(1 - abs(a - b) / (a + b + .0001)))} \cr
  \code{cosine} \tab \code{sum(a * b) / sum(a^2 * sum(b^2))^.5} \cr
  \code{pearson} \tab \code{cor(a, b, method='pearson')} \cr
  \code{kendall} \tab \code{cor(a, b, method='kendall')} \cr
  \code{spearman} \tab \code{cor(a, b, method='spearman')} \cr
  \code{jaccard} \tab \code{sum(a & b) / sum(a | b)} \cr
  \code{kld} \tab \code{p = (p<-table(a,b)/100)/sum(p); q = rowSums(p) \%*\% matrix(colSums(p),1);
    sum(p*log(p/q))}\cr
}}

\item{metric.arg}{a list of arguments to be passed to \code{metric} if \code{metric} is a
function.}

\item{group}{if \code{b} is missing and \code{a} has multiple rows, this will be used to make
comparisons between rows of \code{a}, as modified by \code{agg} and \code{add.mean}.}

\item{agg}{logical; if \code{FALSE}, only the boundary rows between groups will be compared, see
example.}

\item{agg.mean}{logical; if \code{FALSE}, consecutive rows of the same group will be summed.}

\item{square}{logical; if \code{FALSE}, only the lower triangle is returned from a pairwise
comparison.}

\item{mean}{logical; if \code{TRUE}, a single mean for each metric is returned.}

\item{ncores}{sets the number of CPU cores to be used during pairwise comparisons. If not
specified, multiple cores will only be used if \code{nrow(a)} is greater than 1000, in which
case the number of detected cores - 2 will be used.}
}
\description{
Similarity, Distance, and/or Accommodation Metrics
}
\details{
When a function, \code{metric} is called in places of built-in metrics. Two arguments are always
passed to it, generally corresponding to rows of \code{a} in the first position, and \code{b} in
the second. If \code{b} is missing and \code{a} has more than one row, other rows of \code{a}
will be in the second position.

If \code{metric} is a function with compatible first and second position arguments, this can be
entered directly (e.g., \code{metric = energy::dcor}). Otherwise, \code{a} and \code{b} might
have to be reformatted (e.g., \code{metric = function(a,b) entropy::mi.empirical(table(a,b))}).

The function entered as \code{metric} must always return a single numerical value.
}
\examples{
text = c(
  'words of speaker A','more words from speaker A',
  'words from speaker B','more words from speaker B'
)
speaker = c('A','A','B','B')

(dtm = lma_dtm(text))

# by default, consecutive rows from the same group are averaged:
lma_simets(dtm, group=speaker)

# with agg = FALSE, only the rows at the boundary between
# groups (rows 2 and 3 in this case) are used:
lma_simets(dtm, group=speaker, agg=FALSE)
}
