% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.dic}
\alias{read.dic}
\alias{write.dic}
\title{Read/write dictionary files}
\usage{
read.dic(path, cats, type = "asis", as.weighted = FALSE)

write.dic(dict, filename = "custom", type = "asis", as.weighted = FALSE,
  save = TRUE)
}
\arguments{
\item{path}{Path to a file, a name corresponding to a file in \code{getOption('lingmatch.dict.dir')} or
one of the dictionaries available at \href{https://osf.io/y6g5b}{osf.io/y6g5b}, a matrix-like object
to be categorized, or a list to be formatted.}

\item{cats}{A character vector of category names to be returned. All categories are returned by default.}

\item{type}{A character indicating whether and how terms should be altered. Unspecified or matching 'asis'
leaves terms as they are. Other options change wildcards to regular expressions:
\code{'pattern'} (\code{'^[poi]'}) replaces initial asterisks with \code{'\\\\b\\\\w*'},
and terminal asterisks with \code{'\\\\w*\\\\b'}, to match terms within raw text;
for anything else, terms are padded with '^' and '$', then those bounding marks are removed
when an asterisk is present, to match tokenized terms.}

\item{as.weighted}{Logical; if \code{TRUE}, converts to a binary weighted dictionary -- a
data.frame with a "term" column of unique terms, and a column for each category.}

\item{dict}{A list object with names of categories and a vector of their words.}

\item{filename}{The name of the file to be saved.}

\item{save}{Logical; if \code{FALSE}, does not write a file.}
}
\value{
a character vector
}
\description{
Read in or write dictionary files in Comma-Separated Values (.csv; weighted) or
Linguistic Inquiry and Word Count (.dic; non-weighted) format.
}
\examples{
# make a small murder related dictionary
dict = list(
  kill = c('kill*', 'murd*', 'wound*', 'die*'),
  death = c('death*', 'dying', 'die*', 'kill*')
)

# convert it to a weighted format
dict_weighted = read.dic(dict, as.weighted = TRUE)

# categorize it back
read.dic(dict_weighted)

\dontrun{

# save it as a .dic file
write.dic(dict, 'murder')

# read it back in as a list
read.dic('murder.dic')

# read in the Moral Foundations or LUSI dictionaries from urls
moral_dict = read.dic('http://bit.ly/MoralFoundations2')
lusi_dict = read.dic('http://bit.ly/lusi_dict')
}
}
\seealso{
Other Dictionary functions: 
\code{\link{download.dict}()},
\code{\link{lma_patcat}()},
\code{\link{lma_termcat}()},
\code{\link{select.dict}()}
}
\concept{Dictionary functions}
