% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize.lsspace}
\alias{standardize.lsspace}
\title{Standardize a Latent Semantic Space}
\usage{
standardize.lsspace(infile, name, sep = " ", digits = 9,
  dir = getOption("lingmatch.lspace.dir"), outdir = dir, remove = "",
  term_check = "^[a-zA-Z]+$|^['a-zA-Z][a-zA-Z.'\\\\/-]*[a-zA-Z.]$",
  verbose = FALSE)
}
\arguments{
\item{infile}{name of the .rda or plain-text file relative to \code{dir},
e.g., "default.rda" or "glove/glove.6B.300d.txt".}

\item{name}{base name of the reformatted file and term file; e.g., "glove" would result in
\code{glove.dat} and \code{glove_terms.txt} in \code{outdir}.}

\item{sep}{delimiting character between values in each line, e.g., \code{" "} or \code{"\\t"}.
Only applies to plain-text files.}

\item{digits}{number of digits to round values to; default is 9.}

\item{dir}{path to folder containing \code{infile}s; default is \code{getOption('lingmatch.lspace.dir')}.}

\item{outdir}{path to folder in which to save standardized files; defaults to \code{dir}.}

\item{remove}{a string with a regex pattern to be removed from term names (as in \code{gsub(}\code{remove,}
\code{"", term)}); default is \code{""}, which is ignored.}

\item{term_check}{a string with a regex pattern by which to filter terms; i.e., only lines with fully
matched terms are written to the reformatted file. The default attempts to retain only regular words, including
those with dashes, foreword slashes, and periods. Set to an empty string (\code{""}) to write all lines
regardless of term.}

\item{verbose}{logical; if \code{TRUE}, prints the current line number and its term to the console every 1,000 lines.
Only applies to plain-text files.}
}
\description{
Reformat a .rda file which has a matrix with terms as row names, or a plain-text embeddings file
which has a term at the start of each line, and consistent delimiting characters. Plain-text files
are processed line-by-line, so large spaces can be reformatted RAM-conservatively.
}
\examples{
\dontrun{

# from https://sites.google.com/site/fritzgntr/software-resources/semantic_spaces
standardize.lsspace('EN_100k_lsa.rda', '100k_lsa')

# from https://fasttext.cc/docs/en/english-vectors.html
standardize.lsspace('crawl-300d-2M.vec', 'facebook_crawl')
}
}
\seealso{
Other Latent Semantic Space functions: 
\code{\link{download.lsspace}()},
\code{\link{lma_lspace}()},
\code{\link{select.lsspace}()}
}
\concept{Latent Semantic Space functions}
