% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download.lsspace}
\alias{download.lsspace}
\title{Download Latent Semantic Spaces}
\usage{
download.lsspace(space = "default",
  dir = "~/Documents/Latent Semantic Spaces")
}
\arguments{
\item{space}{name of the space you wish to download. Options include \code{'default'} and \code{'tasa'}.}

\item{dir}{desired path to the downloaded space; default is '~/Documents/Latent Semantic Spaces'.}
}
\description{
Downloads the specified semantic space.
}
\details{
Spaces are slightly altered and reprocessed versions of those available at
\url{http://www.lingexp.uni-tuebingen.de/z2/LSAspaces/}.

Spaces can (sometimes) be downloaded directly from \url{https://www.myweb.ttu.edu/miserman/lsspaces/}.
}
