% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download.lsspace}
\alias{download.lsspace}
\title{Download Latent Semantic Spaces}
\usage{
download.lsspace(space = "default", type = "sqlite")
}
\arguments{
\item{space}{name of the space you wish to download. Options include \code{'default'} and \code{'tasa'}.}

\item{type}{the type of file you wish to download, either \code{'rda'} or \code{'sqlite'}. If you plan on
loading the space into memory before use, rdas will be fastest, otherwise use sqlites.}
}
\description{
Downloads the specified semantic space from the to lingmatch github repository to the lingmatch data folder.
}
\details{
Spaces are slightly altered and reprocessed versions of those available at
\url{http://www.lingexp.uni-tuebingen.de/z2/LSAspaces/}.

Spaces can (sometimes) be downloaded directly from \url{https://www.myweb.ttu.edu/miserman/lsspaces/}.
}
