% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lma_patcat}
\alias{lma_patcat}
\title{Categorize raw texts using a pattern-based dictionary}
\usage{
lma_patcat(text, dict, term = "term", category = "category",
  weight = "weight", to.lower = TRUE, to.percent = FALSE,
  bias = NULL, intname = "_intercept", return_dtm = FALSE,
  exclusive = TRUE, boundary = NULL, fixed = TRUE, perl = TRUE,
  ncores = detectCores() - 2)
}
\arguments{
\item{text}{A vector of raw text to be categorized.}

\item{dict}{At least a vector of terms (patterns), usually a matrix-like object with columns for terms,
categories, and weights.}

\item{term, category, weight}{Strings specifying the relevant column names in \code{dict}.}

\item{to.lower}{Logical indicating whether \code{text} should be converted to lower case.}

\item{to.percent}{Logical indicating whether term-counts should be divided by document-counts before
being weighted (defaults to \code{FALSE}).}

\item{bias}{A constant to add to each category after weighting and summing. Can be a vector with names
corresponding to the unique values in \code{dict[,category]}, but is usually extracted from dict based
on the intercept included in each category (defined by \code{intname}).}

\item{intname}{The term representing the intercept (bias) of a category, to be extracted from \code{dict}
and used as \code{bias}.}

\item{return_dtm}{Logical; if \code{TRUE}, only a document-term matrix will be returned, rather than the
weighted, summed, and adjusted category value.}

\item{exclusive}{Logical; if \code{FALSE}, each dictionary term is searched for in the original text.
Otherwise (by default), terms are sorted by length (with longer terms being searched for first), and
matches are removed from the text (avoiding subsequent matched to matched patterns).}

\item{boundary}{A string to add to the beginning and end of each dictionary term. If \code{TRUE},
\code{boundary} will be set to \code{' '}, avoiding pattern matches within words. By default, dictionary
terms are left as entered.}

\item{fixed}{Logical; if \code{FALSE}, patterns can be regular expressions.}

\item{perl}{Logical; passed to \code{\link{strsplit}. Set to \code{FALSE} if \code{fixed} is \code{TRUE}}.}

\item{ncores}{Number of CPU cores to use. Default is number of cores - 2 if \code{text} has more than 100
entries. Otherwise, if not specified, only 1 core will be used.}
}
\description{
Categorize raw texts using a pattern-based dictionary
}
\examples{
# example text
text = c(
  paste(
    "Oh, what youth was! What I had and gave away.",
    "What I took and spent and saw. What I lost. And now? Ruin."
  ),
  paste(
    "God, are you so bored?! You just want what's gone from us all?",
    "I miss the you that was too. I love that you."
  ),
  paste(
    "Tomorrow! Tomorrow--nay, even tonight--you wait, as I am about to change.",
    "Soon I will off to revert. Please wait."
  )
)

# read in the temporal orientation lexicon from the World Well-Being Project
tempori = read.csv('https://wwbp.org/downloads/public_data/temporalOrientationLexicon.csv')

lma_patcat(text, tempori)

}
