% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingmatch.R
\name{lma_termcat}
\alias{lma_termcat}
\title{Document-Term Matrix Categorization}
\usage{
lma_termcat(dtm, dict, term.filter = NULL)
}
\arguments{
\item{dtm}{A matrix with words as column names.}

\item{dict}{A \code{list} object with named character vectors as word lists.}

\item{term.filter}{A regular expression string used to format the text of each term (passed to
\code{\link[base]{gsub}}). For example, if terms are part-of-speech tagged (e.g.,
\code{'a_DT'}), \code{filter='_.*'} would remove the tag.}
}
\description{
Reduces the dimensions of a document-term matrix by dictionary-based categorization.
}
