% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.dic}
\alias{read.dic}
\alias{write.dic}
\title{Read/write LIWC dictionary files}
\usage{
read.dic(path, cats)

write.dic(x, filename = "custom")
}
\arguments{
\item{path}{Path to a .dic file.}

\item{cats}{A character vector of category names to be returned. All categories are returned by default.}

\item{x}{A list object with names of categories and a vector of their words.}

\item{filename}{The name of the file to be saved.}
}
\description{
Read in or write Linguistic Inquiry and Word Count dictionary (.dic) files.
}
\examples{
# make a small murder related dictionary
dict = list(
  kill = c('kill*', 'murd*', 'wound*', 'die*'),
  death = c('death*', 'dying', 'die*', 'kill*')
)

# save it as a .dic file: write.dic(dict,'murder')
# read it back in as a list: read.dic('murder.dic')

# read in the Moral Foundations or LUSI dictionaries from urls
moral_dict = read.dic('http://bit.ly/MoralFoundations')
lusi_dict = read.dic('http://bit.ly/lusi_dict')

}
