% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.dic}
\alias{read.dic}
\alias{write.dic}
\title{Read/write LIWC dictionary files}
\usage{
read.dic(path, cats, to.regex = FALSE)

write.dic(dict, filename = "custom", save = TRUE)
}
\arguments{
\item{path}{Path to a .dic file.}

\item{cats}{A character vector of category names to be returned. All categories are returned by default.}

\item{to.regex}{Logical; if \code{TRUE}, each dictionary entry is converted to regular expression
(starts are marked with ^ and ends with $ when an * is present, and unmatched brackets/parens are escaped).}

\item{dict}{A list object with names of categories and a vector of their words.}

\item{filename}{The name of the file to be saved.}

\item{save}{Logical; if \code{FALSE}, does not write a file.}
}
\value{
a character vector
}
\description{
Read in or write Linguistic Inquiry and Word Count dictionary (.dic) files.
}
\examples{
# make a small murder related dictionary
dict = list(
  kill = c('kill*', 'murd*', 'wound*', 'die*'),
  death = c('death*', 'dying', 'die*', 'kill*')
)

\dontrun{

write.dic(dict, 'murder') # save it as a .dic file
read.dic('murder.dic') # read it back in as a list

# read in the Moral Foundations or LUSI dictionaries from urls
moral_dict = read.dic('http://bit.ly/MoralFoundations2')
lusi_dict = read.dic('http://bit.ly/lusi_dict')
}
}
