% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.folder}
\alias{read.folder}
\title{Process texts in a folder}
\usage{
read.folder(path = NULL, segment = NULL, subdir = FALSE,
  ext = ".txt", fixed = TRUE, segment.size = NULL,
  bysentence = FALSE, reader = readLines, readarg = list(warn =
  FALSE), ncores = parallel::detectCores() - 2)
}
\arguments{
\item{path}{Path to a folder containing files, or a vector of paths to files.}

\item{segment}{Specifies how the text of each file should be segmented. If a number, texts will be broken into
that many segments, each with a roughly equal number of words. If a character, texts will be broken at that character;
for example, a string matching \code{join} will split texts on returns.}

\item{subdir}{If \code{TRUE} files in folders in \code{path} will also be included.}

\item{ext}{The extension of the files you want to read in. '.txt' by default.}

\item{fixed}{If \code{FALSE}, and \code{segment} is a character, \code{segment} will be treated as a regular expression.}

\item{segment.size}{If specified, \code{segment} will be ignored, and texts will be broken into segments containing roughly
\code{segment.size} number of words.}

\item{bysentence}{If \code{TRUE}, and \code{segment} is a number or \code{wordcount} is specified, sentences will be kept
together, rather than being broken across segments.}

\item{reader}{The function used to read files. File paths are always passed as the first argument.
\code{\link[base]{readLines}} by default.}

\item{readarg}{A list of additional arguments to pass to \code{reader}, starting in the second position.}

\item{ncores}{Number of CPU cores to use; defaults to total number of cores minus 2.}
}
\description{
Read in and optionally segment all texts within a folder.
}
