% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select.lspace}
\alias{select.lspace}
\title{Select Latent Semantic Spaces}
\usage{
select.lspace(query = NULL, dir = getOption("lingmatch.lspace.dir"),
  get.map = FALSE, check.md5 = TRUE, mode = "wb")
}
\arguments{
\item{query}{A character matching a space name, or a character vector of terms, used
to select spaces. If length is over 1, \code{get.map} is set to \code{TRUE}.}

\item{dir}{Path to a directory containing \code{lma_term_map.rda} and downloaded spaces; \cr will look in
\code{getOption('lingmatch.lspace.dir')} and \code{'~/Latent Semantic Spaces'} by default.}

\item{get.map}{Logical; if \code{TRUE} and \code{lma_term_map.rda} is not found in
\code{dir}, the term map (\href{https://osf.io/xr7jv}{lma_term_map.rda}) is
downloaded and decompressed.}

\item{check.md5}{Logical; if \code{TRUE} (default), retrieves the MD5 checksum from OSF,
and compares it with that calculated from the downloaded file to check its integrity.}

\item{mode}{Passed to \code{\link{download.file}} when downloading the term map.}
}
\value{
A list with varying entries:
  \itemize{
    \item \strong{\code{info}}: The version of \href{https://osf.io/9yzca}{osf.io/9yzca} stored internally; a
      \code{data.frame}  with spaces as row names, and information about each space in columns:
        \itemize{
          \item \strong{\code{terms}}: number of terms in the space
          \item \strong{\code{corpus}}: corpus(es) on which the space was trained
          \item \strong{\code{model}}: model from which the space was trained
          \item \strong{\code{dimensions}}: number of dimensions in the model (columns of the space)
          \item \strong{\code{model_info}}: some parameter details about the model
          \item \strong{\code{original_max}}: maximum value used to normalize the space; the original
            space would be \code{(vectors *} \code{original_max) /} \code{100}
          \item \strong{\code{osf_dat}}: OSF id for the \code{.dat} files; the URL would be
            https://osf.io/\code{osf_dat}
          \item \strong{\code{osf_terms}}: OSF id for the \code{_terms.txt} files; the URL would be
            https://osf.io/\code{osf_terms}
          \item \strong{\code{wiki}}: link to the wiki for the space
          \item \strong{\code{downloaded}}: path to the \code{.dat} file if downloaded,
            and \code{''} otherwise.
        }
    \item \strong{\code{selected}}: A subset of \code{info} selected by \code{query}.
    \item \strong{\code{term_map}}: If \code{get.map} is \code{TRUE} or \code{lma_term_map.rda} is found in
      \code{dir}, a copy of \href{https://osf.io/xr7jv}{osf.io/xr7jv}, which has space names as
      column names, terms as row names, and indices as values, with 0 indicating the term is not
      present in the associated space.
  }
}
\description{
Retrieve information and links to latent semantic spaces
(sets of word vectors/embeddings) available at \href{https://osf.io/489he}{osf.io/489he},
and optionally download their term mappings (\href{https://osf.io/xr7jv}{osf.io/xr7jv}).
}
\examples{
# just retrieve information about available spaces
spaces = select.lspace()

# retrieve all spaces that used word2vec
w2v_spaces = select.lspace('word2vec')$selected

\dontrun{

# select spaces by terms
select.lspace(c(
  'part-time', 'i/o', "'cause", 'brexit', 'debuffs'
))$selected[, c('terms', 'coverage')]
}
}
\seealso{
Other Latent Semantic Space functions: 
\code{\link{download.lspace}()},
\code{\link{lma_lspace}()},
\code{\link{standardize.lspace}()}
}
\concept{Latent Semantic Space functions}
